Some additional cleanups
