+++ test_visual.py
@@ -1,0 +1,12 @@
+## WORK IN PROGRESS :D
+
+# import pytest
+#
+# from tests.notebook_fixtures import all_notebooks
+#
+#
+# @pytest.mark.pdf
+# @pytest.mark.parametrize("notebook", all_notebooks, indirect=True)
+# @visual_inspection
+# def test_valid_pdf(notebook, remarks_document):
+#     assert remarks_document.is_pdf
