--- tests/in/pdf_with_multiple_added_pages/6c4fa130-6a4a-405e-9726-438ba8bc8238.metadata
@@ -1,14 +1,0 @@
-{
-    "deleted": false,
-    "lastModified": "1690133856920",
-    "lastOpened": "1690133859119",
-    "lastOpenedPage": 0,
-    "metadatamodified": false,
-    "modified": false,
-    "parent": "",
-    "pinned": false,
-    "synced": false,
-    "type": "DocumentType",
-    "version": 0,
-    "visibleName": "pdf_longer"
-}
