--- tests/in/pdf_with_multiple_added_pages/6c4fa130-6a4a-405e-9726-438ba8bc8238.pagedata
@@ -1,3 +1,0 @@
-Blank
-
-
