--- tests/in/v3_annotated_notebook/1f830c35-1e0d-4ba1-80aa-5acc85d8859a.content
@@ -1,74 +1,0 @@
-{
-    "cPages": {
-        "lastOpened": {
-            "timestamp": "0:0",
-            "value": ""
-        },
-        "original": {
-            "timestamp": "0:0",
-            "value": -1
-        },
-        "pages": [
-            {
-                "id": "e19dd33a-5b98-4b03-9ea9-e93a0bb336ad",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "ba"
-                },
-                "scrollTime": {
-                    "timestamp": "1:4",
-                    "value": "2023-01-20T16:30:39Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:4",
-                    "value": 1872
-                }
-            }
-        ],
-        "uuids": [
-            {
-                "first": "a533ab32-689b-5e15-976d-4671496a1fc0",
-                "second": 1
-            }
-        ]
-    },
-    "coverPageNumber": -1,
-    "customZoomCenterX": 0,
-    "customZoomCenterY": 936,
-    "customZoomOrientation": "portrait",
-    "customZoomPageHeight": 1872,
-    "customZoomPageWidth": 1404,
-    "customZoomScale": 1,
-    "documentMetadata": {
-    },
-    "extraMetadata": {
-        "LastBallpointv2Color": "Red",
-        "LastBallpointv2Size": "2",
-        "LastEraserColor": "Black",
-        "LastEraserSize": "2",
-        "LastEraserTool": "Eraser",
-        "LastPaintbrushv2Color": "Gray",
-        "LastPaintbrushv2Size": "2",
-        "LastPen": "Ballpointv2",
-        "LastTool": "Ballpointv2"
-    },
-    "fileType": "notebook",
-    "fontName": "",
-    "formatVersion": 2,
-    "lineHeight": -1,
-    "margins": 125,
-    "orientation": "portrait",
-    "pageCount": 1,
-    "pageTags": [
-    ],
-    "sizeInBytes": "7859",
-    "tags": [
-    ],
-    "textAlignment": "justify",
-    "textScale": 1,
-    "zoomMode": "bestFit"
-}
