--- tests/in/v3_annotated_notebook/1f830c35-1e0d-4ba1-80aa-5acc85d8859a.metadata
@@ -1,9 +1,0 @@
-{
-    "lastModified": "1674232176508",
-    "lastOpened": "1674232142624",
-    "lastOpenedPage": 0,
-    "parent": "",
-    "pinned": false,
-    "type": "DocumentType",
-    "visibleName": "V3 sample"
-}
