--- tests/in/v3_annotated_notebook/1f830c35-1e0d-4ba1-80aa-5acc85d8859a.pagedata
@@ -1,1 +1,0 @@
-P Dots S
