--- tests/in/v3_coords_notebook/84099c77-bd7b-4a0d-9e06-a11c89e2bca4.content
@@ -1,245 +1,0 @@
-{
-    "cPages": {
-        "lastOpened": {
-            "timestamp": "1:22",
-            "value": "0c220082-a11a-4c53-8d60-ab9217aef034"
-        },
-        "original": {
-            "timestamp": "0:0",
-            "value": -1
-        },
-        "pages": [
-            {
-                "id": "f397734a-15bf-4a69-894d-81b0e2ba50d1",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "ba"
-                },
-                "scrollTime": {
-                    "timestamp": "1:5",
-                    "value": "2023-01-20T18:11:13Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:5",
-                    "value": 1872
-                }
-            },
-            {
-                "id": "c5ae6db6-aa4a-4744-a43a-2f1928796449",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bb"
-                },
-                "scrollTime": {
-                    "timestamp": "1:3",
-                    "value": "2023-01-20T20:48:18Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:3",
-                    "value": 1872
-                }
-            },
-            {
-                "id": "34b966ca-f36d-4d44-91e9-8067580c728d",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bc"
-                },
-                "scrollTime": {
-                    "timestamp": "1:4",
-                    "value": "2023-01-20T20:48:21Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:4",
-                    "value": 1872
-                }
-            },
-            {
-                "id": "199fb186-cd1b-4fd6-94c1-7689a7e42605",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bd"
-                },
-                "scrollTime": {
-                    "timestamp": "1:3",
-                    "value": "2023-01-20T21:06:43Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:3",
-                    "value": 1872
-                }
-            },
-            {
-                "id": "381e5d1b-d228-483e-9af1-54573272fec6",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "be"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                }
-            },
-            {
-                "id": "0c220082-a11a-4c53-8d60-ab9217aef034",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bf"
-                },
-                "scrollTime": {
-                    "timestamp": "1:1",
-                    "value": "2023-01-22T11:44:08Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:1",
-                    "value": 1872
-                }
-            },
-            {
-                "deleted": {
-                    "timestamp": "1:1",
-                    "value": 1
-                },
-                "id": "89e383a5-b9b6-4568-814b-9023001dcdc4",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bb"
-                },
-                "scrollTime": {
-                    "timestamp": "1:2",
-                    "value": "2023-01-20T18:11:22Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:2",
-                    "value": 1872
-                }
-            },
-            {
-                "deleted": {
-                    "timestamp": "1:1",
-                    "value": 1
-                },
-                "id": "9d492eda-595f-40cb-92e9-034cf44f9b9c",
-                "idx": {
-                    "timestamp": "1:1",
-                    "value": "bb"
-                },
-                "scrollTime": {
-                    "timestamp": "1:3",
-                    "value": "2023-01-20T18:10:53Z"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "P Dots S"
-                },
-                "verticalScroll": {
-                    "timestamp": "1:3",
-                    "value": 1872
-                }
-            }
-        ],
-        "uuids": [
-            {
-                "first": "a533ab32-689b-5e15-976d-4671496a1fc0",
-                "second": 1
-            }
-        ]
-    },
-    "coverPageNumber": -1,
-    "customZoomCenterX": 0,
-    "customZoomCenterY": 936,
-    "customZoomOrientation": "portrait",
-    "customZoomPageHeight": 1872,
-    "customZoomPageWidth": 1404,
-    "customZoomScale": 1,
-    "documentMetadata": {
-    },
-    "extraMetadata": {
-        "LastBallpointColor": "Black",
-        "LastBallpointSize": "2",
-        "LastBallpointv2Color": "Red",
-        "LastBallpointv2Size": "2",
-        "LastCalligraphyColor": "Black",
-        "LastCalligraphySize": "2",
-        "LastClearPageColor": "Black",
-        "LastClearPageSize": "2",
-        "LastEraseSectionColor": "Black",
-        "LastEraseSectionSize": "2",
-        "LastEraserColor": "Black",
-        "LastEraserSize": "2",
-        "LastEraserTool": "Eraser",
-        "LastFinelinerColor": "Black",
-        "LastFinelinerSize": "2",
-        "LastFinelinerv2Color": "Black",
-        "LastFinelinerv2Size": "2",
-        "LastHighlighterColor": "Black",
-        "LastHighlighterSize": "2",
-        "LastHighlighterv2Color": "Black",
-        "LastHighlighterv2Size": "2",
-        "LastMarkerColor": "Black",
-        "LastMarkerSize": "2",
-        "LastMarkerv2Color": "Black",
-        "LastMarkerv2Size": "2",
-        "LastPaintbrushColor": "Black",
-        "LastPaintbrushSize": "2",
-        "LastPaintbrushv2Color": "Gray",
-        "LastPaintbrushv2Size": "2",
-        "LastPen": "Ballpointv2",
-        "LastPencilColor": "Black",
-        "LastPencilSize": "2",
-        "LastPencilv2Color": "Black",
-        "LastPencilv2Size": "2",
-        "LastReservedPenColor": "Black",
-        "LastReservedPenSize": "2",
-        "LastSelectionToolColor": "Black",
-        "LastSelectionToolSize": "2",
-        "LastSharpPencilColor": "Black",
-        "LastSharpPencilSize": "2",
-        "LastSharpPencilv2Color": "Black",
-        "LastSharpPencilv2Size": "2",
-        "LastSolidPenColor": "Black",
-        "LastSolidPenSize": "2",
-        "LastTool": "Ballpointv2",
-        "LastZoomToolColor": "Black",
-        "LastZoomToolSize": "2"
-    },
-    "fileType": "notebook",
-    "fontName": "",
-    "formatVersion": 2,
-    "lineHeight": -1,
-    "margins": 125,
-    "orientation": "portrait",
-    "pageCount": 6,
-    "pageTags": [
-    ],
-    "sizeInBytes": "112925",
-    "tags": [
-    ],
-    "textAlignment": "justify",
-    "textScale": 1,
-    "zoomMode": "bestFit"
-}
