--- tests/in/v3_coords_notebook/84099c77-bd7b-4a0d-9e06-a11c89e2bca4.metadata
@@ -1,9 +1,0 @@
-{
-    "lastModified": "1674238570727",
-    "lastOpened": "1674386799766",
-    "lastOpenedPage": 5,
-    "parent": "",
-    "pinned": false,
-    "type": "DocumentType",
-    "visibleName": "V3coords"
-}
