--- tests/in/v3_coords_notebook/84099c77-bd7b-4a0d-9e06-a11c89e2bca4.pagedata
@@ -1,1 +1,0 @@
-P Dots S
