--- tests/in/v3_simple_notebook/f8000ca5-cc1f-42eb-bf9e-3ca1145c38d6.metadata
@@ -1,14 +1,0 @@
-{
-    "deleted": false,
-    "lastModified": "1671633025409",
-    "lastOpened": "1671633025878",
-    "lastOpenedPage": 0,
-    "metadatamodified": false,
-    "modified": true,
-    "parent": "",
-    "pinned": false,
-    "synced": false,
-    "type": "DocumentType",
-    "version": 0,
-    "visibleName": "Beta3nb"
-}
