--- tests/in/v3_simple_notebook/f8000ca5-cc1f-42eb-bf9e-3ca1145c38d6.pagedata
@@ -1,1 +1,0 @@
-P Dots S
