--- tests/in/v3_typed_text/08ac1700-d8c4-4350-938c-ac0eaa00916c.content
@@ -1,68 +1,0 @@
-{
-    "cPages": {
-        "lastOpened": {
-            "timestamp": "1:3",
-            "value": "f79e7612-cd68-4140-9665-d4d33a0110ab"
-        },
-        "original": {
-            "timestamp": "0:0",
-            "value": -1
-        },
-        "pages": [
-            {
-                "id": "f79e7612-cd68-4140-9665-d4d33a0110ab",
-                "idx": {
-                    "timestamp": "1:2",
-                    "value": "ba"
-                },
-                "template": {
-                    "timestamp": "1:1",
-                    "value": "Blank"
-                }
-            },
-            {
-                "id": "d0b0f317-a41e-410f-a0d4-25d5e5711fb4",
-                "idx": {
-                    "timestamp": "1:2",
-                    "value": "bb"
-                },
-                "template": {
-                    "timestamp": "1:2",
-                    "value": "Blank"
-                }
-            }
-        ],
-        "uuids": [
-            {
-                "first": "4f498988-8737-53cf-890e-4b96d733c02e",
-                "second": 1
-            }
-        ]
-    },
-    "coverPageNumber": -1,
-    "customZoomCenterX": 0,
-    "customZoomCenterY": 936,
-    "customZoomOrientation": "portrait",
-    "customZoomPageHeight": 1872,
-    "customZoomPageWidth": 1404,
-    "customZoomScale": 1,
-    "documentMetadata": {
-    },
-    "extraMetadata": {
-    },
-    "fileType": "notebook",
-    "fontName": "",
-    "formatVersion": 2,
-    "lineHeight": -1,
-    "margins": 125,
-    "orientation": "portrait",
-    "pageCount": 2,
-    "pageTags": [
-    ],
-    "sizeInBytes": "131212",
-    "tags": [
-    ],
-    "textAlignment": "justify",
-    "textScale": 1,
-    "zoomMode": "bestFit"
-}
