--- tests/in/v3_typed_text/08ac1700-d8c4-4350-938c-ac0eaa00916c.metadata
@@ -1,10 +1,0 @@
-{
-    "createdTime": "1727914856867",
-    "lastModified": "1727914946612",
-    "lastOpened": "1727914966189",
-    "lastOpenedPage": 0,
-    "parent": "",
-    "pinned": false,
-    "type": "DocumentType",
-    "visibleName": "Github Notebook"
-}
